requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
loguru==0.6.0
MarkupSafe==2.1.1
mergedict==1.0.0
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
